move custom hook to its own file
